+++ CHANGELOG.md
@@ -14,6 +14,10 @@ We use the following categories for changes:
 
 ## [Unreleased]
 
+### Added
+- Support to create and delete metric downsampling based on metric types [#538]
+- Support for maintenance of metric downsampling [#554]
+
 ## [0.8.0 - 2023-01-05]
 
 ### Changed
@@ -39,8 +43,6 @@ We use the following categories for changes:
 
 ### Fixed
 - Column conflict when creating a metric view with a label called `series` [#559]
-- Support to create and delete metric rollups based on metric types [#538]
-- Support for maintenance of Metric-rollups and Caggs for downsampling [#554]
 
 ## [0.7.0 - 2022-10-03]